src/main/java/com/kakao/domain/ladder/Ladder.java
package com.kakao.domain.ladder;

import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

public class Ladder {

    private final List<Line> lines;

    private Ladder(List<Line> lines) {
        this.lines = Collections.unmodifiableList(lines);
    }

    public static Ladder create(int width, int height) {
        List<Line> lines = new ArrayList<>();
        for (int i = 0; i < height; i++) {
            lines.add(Line.drawLine(width));
        }
        return new Ladder(lines);
    }

    public List<Line> getLines() {
        return lines;
    }
}